+++ Changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming verstion
+* [TASK] Make extension compatible with TYPO3 13
+
 ## Version 1.1.0 - 2024-10-04
 * [TASK] Set related_links to 0
 * [TASK] Add Changelog